+++ src/julia/elements/book1/def_006_surfextrem.jl
@@ -66,31 +66,49 @@ function hide_edge_and_collapse(
     corner::Vector{Float32})
 
     OdinJuliaBridge.hide_point(state_ptr, host_id)
-    OdinJuliaBridge.set_point_position(state_ptr, joint1_id, corner[1], corner[2], corner[3])
-    OdinJuliaBridge.set_point_position(state_ptr, joint2_id, corner[1], corner[2], corner[3])
+    OdinJuliaBridge.set_point_position(
+        state_ptr, joint1_id, corner[1], corner[2], corner[3])
+    OdinJuliaBridge.set_point_position(
+        state_ptr, joint2_id, corner[1], corner[2], corner[3])
 end
 
 function reset_cycle_state(state_ptr::Ptr{Cvoid})
-    edge1_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge1HostId))
-    edge1_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge1Joint1Id))
-    edge1_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge1Joint2Id))
+    edge1_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge1HostId))
+    edge1_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge1Joint1Id))
+    edge1_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge1Joint2Id))
 
-    edge2_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge2HostId))
-    edge2_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge2Joint1Id))
-    edge2_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge2Joint2Id))
+    edge2_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge2HostId))
+    edge2_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge2Joint1Id))
+    edge2_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge2Joint2Id))
 
-    edge3_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge3HostId))
-    edge3_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge3Joint1Id))
-    edge3_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge3Joint2Id))
+    edge3_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge3HostId))
+    edge3_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge3Joint1Id))
+    edge3_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge3Joint2Id))
 
-    edge4_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge4HostId))
-    edge4_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge4Joint1Id))
-    edge4_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge4Joint2Id))
+    edge4_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge4HostId))
+    edge4_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge4Joint1Id))
+    edge4_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge4Joint2Id))
 
-    hide_edge_and_collapse(state_ptr, edge1_host_id, edge1_joint1_id, edge1_joint2_id, Corner1)
-    hide_edge_and_collapse(state_ptr, edge2_host_id, edge2_joint1_id, edge2_joint2_id, Corner2)
-    hide_edge_and_collapse(state_ptr, edge3_host_id, edge3_joint1_id, edge3_joint2_id, Corner3)
-    hide_edge_and_collapse(state_ptr, edge4_host_id, edge4_joint1_id, edge4_joint2_id, Corner4)
+    hide_edge_and_collapse(
+        state_ptr, edge1_host_id, edge1_joint1_id, edge1_joint2_id, Corner1)
+    hide_edge_and_collapse(
+        state_ptr, edge2_host_id, edge2_joint1_id, edge2_joint2_id, Corner2)
+    hide_edge_and_collapse(
+        state_ptr, edge3_host_id, edge3_joint1_id, edge3_joint2_id, Corner3)
+    hide_edge_and_collapse(
+        state_ptr, edge4_host_id, edge4_joint1_id, edge4_joint2_id, Corner4)
 
     OdinJuliaBridge.set_animation_meta(state_ptr, MetaPhase, PhaseDescend)
     OdinJuliaBridge.set_animation_meta(state_ptr, MetaTimer, 0f0)
@@ -123,21 +141,21 @@ function initialize(state_ptr::Ptr{Cvoid})
         Corner4[1], Corner4[2], Corner4[3],
         LineColor4, 0f0)
 
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge1HostId, Float32(edge1.host_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge1Joint1Id, Float32(edge1.joint1_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge1Joint2Id, Float32(edge1.joint2_id))
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge1HostId, edge1.host_id)
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge1Joint1Id, edge1.joint1_id)
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge1Joint2Id, edge1.joint2_id)
 
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge2HostId, Float32(edge2.host_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge2Joint1Id, Float32(edge2.joint1_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge2Joint2Id, Float32(edge2.joint2_id))
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge2HostId, edge2.host_id)
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge2Joint1Id, edge2.joint1_id)
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge2Joint2Id, edge2.joint2_id)
 
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge3HostId, Float32(edge3.host_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge3Joint1Id, Float32(edge3.joint1_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge3Joint2Id, Float32(edge3.joint2_id))
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge3HostId, edge3.host_id)
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge3Joint1Id, edge3.joint1_id)
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge3Joint2Id, edge3.joint2_id)
 
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge4HostId, Float32(edge4.host_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge4Joint1Id, Float32(edge4.joint1_id))
-    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge4Joint2Id, Float32(edge4.joint2_id))
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge4HostId, edge4.host_id)
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge4Joint1Id, edge4.joint1_id)
+    OdinJuliaBridge.set_animation_meta(state_ptr, MetaEdge4Joint2Id, edge4.joint2_id)
 
     reset_cycle_state(state_ptr)
 end
@@ -146,21 +164,33 @@ function clean(state_ptr::Ptr{Cvoid})
 end
 
 function loop(state_ptr::Ptr{Cvoid}, dt::Float32)
-    edge1_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge1HostId))
-    edge1_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge1Joint1Id))
-    edge1_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge1Joint2Id))
+    edge1_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge1HostId))
+    edge1_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge1Joint1Id))
+    edge1_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge1Joint2Id))
 
-    edge2_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge2HostId))
-    edge2_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge2Joint1Id))
-    edge2_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge2Joint2Id))
+    edge2_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge2HostId))
+    edge2_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge2Joint1Id))
+    edge2_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge2Joint2Id))
 
-    edge3_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge3HostId))
-    edge3_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge3Joint1Id))
-    edge3_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge3Joint2Id))
+    edge3_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge3HostId))
+    edge3_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge3Joint1Id))
+    edge3_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge3Joint2Id))
 
-    edge4_host_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge4HostId))
-    edge4_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge4Joint1Id))
-    edge4_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(state_ptr, MetaEdge4Joint2Id))
+    edge4_host_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge4HostId))
+    edge4_joint1_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge4Joint1Id))
+    edge4_joint2_id = Integer(OdinJuliaBridge.get_animation_meta(
+        state_ptr, MetaEdge4Joint2Id))
 
     if edge1_host_id < 0
         return
@@ -229,10 +259,14 @@ function loop(state_ptr::Ptr{Cvoid}, dt::Float32)
             timer = 0f0
         end
     elseif phase == PhaseHideLines
-        hide_edge_and_collapse(state_ptr, edge1_host_id, edge1_joint1_id, edge1_joint2_id, Corner1)
-        hide_edge_and_collapse(state_ptr, edge2_host_id, edge2_joint1_id, edge2_joint2_id, Corner2)
-        hide_edge_and_collapse(state_ptr, edge3_host_id, edge3_joint1_id, edge3_joint2_id, Corner3)
-        hide_edge_and_collapse(state_ptr, edge4_host_id, edge4_joint1_id, edge4_joint2_id, Corner4)
+        hide_edge_and_collapse(
+            state_ptr, edge1_host_id, edge1_joint1_id, edge1_joint2_id, Corner1)
+        hide_edge_and_collapse(
+            state_ptr, edge2_host_id, edge2_joint1_id, edge2_joint2_id, Corner2)
+        hide_edge_and_collapse(
+            state_ptr, edge3_host_id, edge3_joint1_id, edge3_joint2_id, Corner3)
+        hide_edge_and_collapse(
+            state_ptr, edge4_host_id, edge4_joint1_id, edge4_joint2_id, Corner4)
 
         timer += dt
         if timer >= HidePauseDuration
